+++ src/app/app.module.ts
@@ -1,9 +1,13 @@
 import { CommonModule } from '@angular/common'; 
 import { HTTP_INTERCEPTORS, HttpClientModule } from '@angular/common/http';
 import { ErrorHandler, NgModule } from '@angular/core';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
+import { MatChipsModule } from '@angular/material/chips';
+import { MatIconModule } from '@angular/material/icon';
+import { MatSelectModule } from '@angular/material/select';
 import { MatSortModule } from '@angular/material/sort';
 import { MatTableModule } from '@angular/material/table';
 
@@ -15,6 +19,8 @@ import { PageNotFoundComponent } from '../components/page-not-found/page-not-fou
 import { GlobalErrorHandlerService } from '../services/global-error-handler.service';
 import { HttpErrorInterceptor } from '../interceptors/http-error-interceptor';
 import { TableComponent } from '../elements/table/table.component';
+import { ChipsComponent } from '../elements/chips/chips.component';
+import { MatAutocompleteModule } from '@angular/material/autocomplete';
 
 @NgModule({
   declarations: [
@@ -22,7 +28,8 @@ import { TableComponent } from '../elements/table/table.component';
     DashboardComponent,
     ErrorComponent,
     PageNotFoundComponent,
-    TableComponent
+    TableComponent,
+    ChipsComponent
   ],
   imports: [
     CommonModule,
@@ -30,14 +37,20 @@ import { TableComponent } from '../elements/table/table.component';
     AppRoutingModule,
     BrowserAnimationsModule,
     HttpClientModule,
+    FormsModule,
+    ReactiveFormsModule,
 
     // Material
+    MatAutocompleteModule,
+    MatChipsModule,
+    MatIconModule,
+    MatSelectModule,
     MatSortModule,
     MatTableModule
   ],
   providers: [
     { provide: HTTP_INTERCEPTORS, useClass: HttpErrorInterceptor, multi: true },
-    { provide: ErrorHandler, useClass: GlobalErrorHandlerService },
+    // { provide: ErrorHandler, useClass: GlobalErrorHandlerService },
   ],
   bootstrap: [AppComponent]
 })